requirements.txt
# This is a list of pypi packages to be installed into virtualenv. Alternatively,
# you can install these as RPMs instead of pypi packages.

fedora_messaging
Flask
Flask-RESTful!=0.3.6
Flask-SQLAlchemy
SQLAlchemy
gssapi
flask-oidc
# packages for the unit tests
pytest >= 2.4.2
mock
stomp.py
Flask-Migrate
six

# Documentation requirements
sphinx
sphinxcontrib-httpdomain

# CLI
click
configparser
openidc-client
requests-gssapi

# Database
psycopg2-binary

# Monitoring
prometheus_client
